+++ apps/app/components/fields/MultipleAmountField.tsx
@@ -4,7 +4,8 @@ import { Input } from '@poveroh/ui/components/input'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@poveroh/ui/components/select'
 import { Button } from '@poveroh/ui/components/button'
 import { Checkbox } from '@poveroh/ui/components/checkbox'
-import { IBankAccount, MultipleAmountFieldProps } from '@poveroh/types'
+import { MultipleAmountFieldProps } from '@/types'
+import { IBankAccount } from '@poveroh/types'
 import { BrandIcon } from '@/components/icon/BrandIcon'
 import { Plus, Trash2, Split, Merge } from 'lucide-react'
 import { BankAccountField } from './BankAccountField'
