+++ apps/app/components/fields/PasswordField.tsx
@@ -1,7 +1,7 @@
 import { FieldValues, Path } from 'react-hook-form'
 import { FormControl, FormField, FormItem, FormLabel, FormMessage, FormDescription } from '@poveroh/ui/components/form'
 import PasswordInput from '@poveroh/ui/components/password'
-import { DescriptiveFieldProps, AutoCompleteFieldProps } from '@poveroh/types'
+import { DescriptiveFieldProps, AutoCompleteFieldProps } from '@/types'
 
 type PasswordFormFieldProps<T extends FieldValues = FieldValues> = DescriptiveFieldProps<T> & AutoCompleteFieldProps<T>
