+++ apps/app/components/form/transactions/IncomeForm.tsx
@@ -9,8 +9,6 @@ import { useTranslations } from 'next-intl'
 import {
     Currencies,
     currencyCatalog,
-    FormProps,
-    FormRef,
     IBankAccount,
     ICategory,
     IItem,
@@ -35,7 +33,7 @@ import { useError } from '@/hooks/useError'
 import { useCategory } from '@/hooks/useCategory'
 import { useBankAccount } from '@/hooks/useBankAccount'
 import logger from '@/lib/logger'
-import { amountSchema } from '@/types/form'
+import { amountSchema, FormProps, FormRef } from '@/types/form'
 
 export const IncomeForm = forwardRef<FormRef, FormProps>((props: FormProps, ref) => {
     const t = useTranslations()
