+++ packages/types/src/lib/input.ts
@@ -1,0 +1,1 @@
+export type InputVariantStyle = 'outlined' | 'contained'
